+++ dd-java-agent/instrumentation/vertx-mysql-client-3.9/build.gradle
@@ -8,11 +8,6 @@ muzzle {
     versions = '[3.9.0,4)'
     assertInverse = true
   }
-  fail {
-    group = 'io.vertx'
-    module = 'vertx-pg-client'
-    versions = '[,]'
-  }
 }
 
 addTestSuiteForDir('latestDepTest', 'test')
@@ -28,7 +23,7 @@ dependencies {
   compileOnly group: 'io.vertx', name: 'vertx-mysql-client', version: '3.9.0'
 
   testImplementation group: 'io.vertx', name: 'vertx-mysql-client', version: '3.9.0'
-  testImplementation project(':dd-java-agent:instrumentation:vertx-sql-common')
+  testRuntimeOnly project(':dd-java-agent:instrumentation:vertx-sql-client-3.9')
 
   // This is needed for the test container to start
   testImplementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.23'
