+++ dd-java-agent/instrumentation/vertx-mysql-client-4.0/build.gradle
@@ -23,14 +23,15 @@ dependencies {
   compileOnly group: 'io.vertx', name: 'vertx-mysql-client', version: '4.0.0'
 
   testImplementation group: 'io.vertx', name: 'vertx-mysql-client', version: '4.0.0'
-  testImplementation project(':dd-java-agent:instrumentation:vertx-sql-common')
+
+  testRuntimeOnly project(':dd-java-agent:instrumentation:vertx-sql-client-3.9')
+  testRuntimeOnly project(':dd-java-agent:instrumentation:vertx-mysql-client-4.4.2')
 
   // This is needed for the test container to start
   testImplementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.23'
   testImplementation group: 'org.testcontainers', name: 'mysql', version: libs.versions.testcontainers.get()
 
   latestDepTestImplementation group: 'io.vertx', name: 'vertx-mysql-client', version: '4.+'
-  latestDepTestImplementation project(':dd-java-agent:instrumentation:vertx-mysql-client-4.4.2')
 }
 
 tasks.withType(Test).configureEach {
