+++ dd-java-agent/instrumentation/vertx-sql-client-3.9/build.gradle
@@ -1,0 +1,13 @@
+apply from: "$rootDir/gradle/java.gradle"
+
+muzzle {
+  pass {
+    group = "io.vertx"
+    module = "vertx-sql-client"
+    versions = "[3.9.0,)"
+  }
+}
+
+dependencies {
+  compileOnly group: 'io.vertx', name: 'vertx-sql-client', version: '3.9.0'
+}
